add drink to list of drinks

+++ app/src/main/java/com/jankuv/alcometer/Shot.kt
@@ -1,7 +1,7 @@
 package com.jankuv.alcometer
 
-class Shot(var drink: String, var volume: Int, var percentage: Int) {
-    fun alcohol_content(): Int {
+class Shot(var drink: String, var volume: Double, var percentage: Int) {
+    fun alcohol_content(): Double {
         return volume * percentage
     }
